refactor: 抽离 ellipsis 公共方法
