+++ src/util/label.ts
@@ -1,6 +1,7 @@
 import { IElement } from '@antv/g-base';
 import { each } from '@antv/util';
 
+/** 获取最长的 label */
 export function getMaxLabelWidth(labels: IElement[]) {
   let max = 0;
   each(labels, (label) => {
@@ -11,4 +12,15 @@ export function getMaxLabelWidth(labels: IElement[]) {
     }
   });
   return max;
+}
+
+/** 获取label长度 */
+export function getLabelLength(isVertical: boolean, label) {
+    const bbox = label.getCanvasBBox();
+    return isVertical ? bbox.width : bbox.height;
+}
+
+/* label长度是否超过约束值 */
+export function testLabel(label: IElement, limitLength: number): boolean {
+    return label.getBBox().width < limitLength;
 }
